Refactor classes into files and setup validators and tests
