+++ res/family_validation.ged
@@ -1,0 +1,8 @@
+0 I01 INDI
+1 NAME Morgan /Freeman/
+1 SEX M
+1 BIRT
+2 DATE 01 JAN 1850
+1 DEAT Y
+2 DATE 02 JAN 2001
+0 TRLR
