+++ res/test_family.ged
@@ -1,0 +1,92 @@
+0 NOTE https://github.com/yayacodes/GEDCOM-PACY
+0 HEAD
+1 SOUR Family Echo
+2 WWW http://www.familyecho.com/
+1 FILE Family 2
+1 DATE 25 JAN 2020
+1 DEST ANSTFILE
+1 GEDC
+2 VERS 5.5.1
+2 FORM LINEAGE-LINKED
+1 SUBM @I2@
+2 NAME Me
+1 SUBN
+1 CHAR UTF-8
+0 @I1@ INDI
+1 NAME Patty
+2 GIVN Patty
+1 SEX F
+1 BIRT
+2 DATE 9 JUN 1972
+1 FAMS @F1@
+1 FAMS @F2@
+1 FAMC @F3@
+0 @I2@ INDI
+1 NAME Me
+2 GIVN Me
+1 SEX M
+1 BIRT
+2 DATE 18 JUN 2009
+1 FAMC @F1@
+0 @I3@ INDI
+1 NAME Chuck
+2 GIVN Chuck
+1 SEX M
+1 BIRT
+2 DATE 10 AUG 1969
+1 DEAT Y
+2 DATE 7 JUN 2004
+1 FAMS @F1@
+0 @I4@ INDI
+1 NAME Dalton
+2 GIVN Dalton
+1 SEX M
+1 BIRT
+2 DATE 5 FEB 1978
+1 FAMS @F2@
+0 @I5@ INDI
+1 NAME Avery
+2 GIVN Avery
+1 SEX F
+1 BIRT
+2 DATE 15 AUG 2009
+1 FAMC @F2@
+0 @I6@ INDI
+1 NAME John
+2 GIVN John
+1 SEX M
+1 BIRT
+2 DATE 15 MAY 1935
+1 FAMS @F3@
+0 @I7@ INDI
+1 NAME Dorothy
+2 GIVN Dorothy
+1 SEX F
+1 BIRT
+2 DATE 20 OCT 1935
+1 FAMS @F3@
+0 @F1@ FAM
+1 HUSB @I3@
+1 WIFE @I1@
+1 CHIL @I2@
+1 MARR
+2 DATE 12 AUG 1994
+1 EVEN
+2 TYPE Ending
+2 DATE 7 JUN 2004
+1 _CURRENT N
+0 @F2@ FAM
+1 HUSB @I4@
+1 WIFE @I1@
+1 CHIL @I5@
+1 MARR
+2 DATE 10 FEB 2007
+1 EVEN
+2 TYPE Ending
+1 _CURRENT N
+0 @F3@ FAM
+1 HUSB @I6@
+1 WIFE @I7@
+1 CHIL @I1@
+1 _CURRENT Y
+0 TRLR
