--- validation.py
@@ -1,9 +1,0 @@
-# Validation methods go here
-def ageValidation(gedcom):
-  # return ['Individual %01 is too old', 'Indi 02 is less than 0']
-  pass
-
-def deathValidation(gedcom):
-  pass
-
-allValidators = [ageValidation, deathValidation]
